+++ crates/codexi/src/logic/balance/mod.rs
@@ -3,5 +3,5 @@
 mod dto;
 mod model;
 
-pub use dto::{AccountBalanceItem, BalanceItem, CodexiBalanceEntry};
-pub use model::{AccountBalance, Balance, CodexiBalance};
+pub use dto::BalanceItem;
+pub use model::Balance;